+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/utils/IOHelpers.java
@@ -15,6 +15,10 @@
  */
 package io.fabric8.kubernetes.client.utils;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -53,5 +57,24 @@ public class IOHelpers {
             }
         }
     }
+
+  public static boolean isJSONValid(String json) throws IOException {
+    boolean valid = true;
+    try{
+      ObjectMapper objectMapper = new ObjectMapper();
+      objectMapper.readTree(json);
+    } catch(JsonProcessingException e){
+      valid = false;
+    }
+    return valid;
+  }
+
+  public static String convertYamlToJson(String yaml) throws IOException {
+    ObjectMapper yamlReader = new ObjectMapper(new YAMLFactory());
+    Object obj = yamlReader.readValue(yaml, Object.class);
+
+    ObjectMapper jsonWriter = new ObjectMapper();
+    return jsonWriter.writeValueAsString(obj);
+  }
 
 }