+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/internal/CertUtilsTest.java
@@ -140,16 +140,6 @@ public class CertUtilsTest {
     verifyFabric8InStore(trustStore);
   }
 
-  @Test
-  public void testECKeyLoad() throws InvalidKeySpecException, CertificateException, NoSuchAlgorithmException, KeyStoreException, IOException {
-    String privateKeyPath = Utils.filePath(getClass().getResource("/ssl/fabric8-ec.key"));
-    String certPath = Utils.filePath(getClass().getResource("/ssl/fabric8-ec.cert"));
-
-    KeyStore trustStore =
-      CertUtils.createKeyStore(null, certPath, null, privateKeyPath, "EC", "foo", null, null);
-
-    assertEquals(1, trustStore.size());
-  }
 
   private void verifyFabric8InStore(KeyStore trustStore)
     throws KeyStoreException, CertificateException, NoSuchAlgorithmException, IOException {