+++ CHANGELOG.md
@@ -6,8 +6,30 @@
 
     * Fix #1297: wrong result produced when exec in used and params contains '&'. Url string not encoded properly.
     * Fix #1449: System.currentTimeMillis() replaced with System.nanoTime()
+    * Fix #1495: avoid runtime dependency on Commons Collections
 
-#### 4.1-SNAPSHOT
+#### 4.2.1
+
+  Bugs
+    
+    * Fix #1297: wrong result produced when exec in used and params contains '&'. Url string not encoded properly.
+    * Fix #1473: Use correct plural form in OpenshiftRole
+    * Fix #1480: The kubernetes-client is not optionally depending on bouncycastle.
+    * Fix #1490: Resource could not be loaded
+    * Fix #1468: Taking labels into consideration when comparing resources for equality.
+  
+  Improvements
+  
+    * Fix #1455: Use SubjectAccessReview and LocalSubjectAccessReview in kubernetes client using subjectAccessReviewAuth()
+  
+  Dependency Upgrade
+  
+
+  New Feature
+  
+    * First Draft of Custom Resource Improvements (#1472)
+
+#### 4.2.0
 
   Bugs